Replace VerticalNav with tab group on common errors page
Switch from VerticalNav sidebar to TabButtons for severity filtering.
Remove layout reset (@) since the page no longer needs a custom layout.
Add WorkflowCommonErrors snippet to workflow run layout.
